Implemented Indexer as interface
