+++ src/main/java/org/javarosa/xpath/eval/IndexerCreator.java
@@ -1,0 +1,10 @@
+package org.javarosa.xpath.eval;
+
+import org.javarosa.core.model.instance.TreeReference;
+
+public interface IndexerCreator {
+
+    Indexer getIndexer(IndexerType indexType, TreeReference expressionRef, TreeReference resultRef);
+
+    Indexer getIndexer(IndexerType indexType, TreeReference expressionRef, TreeReference resultRef, PredicateStep[] predicateSteps);
+}
